Adding the test to a different file
